functions/requirements.txt
firebase-functions==0.1.0
firebase-admin==6.5.0
google-cloud-secret-manager>=2.20.0
requests==2.31.0
